requirements-docker.txt
ogb
pykeops
clearml==1.3.0
einops==0.4.0
hydra-core==1.1.1
numpy==1.21.4
omegaconf==2.1.1
pytorch_lightning==1.5.9
boto3